made changes to search structure ready for server api input
