+++ src/components/Calendar/CalendarEvents.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+const CalendarEvents = (props) => {
+    return (
+        <div>
+        <h1>Event Name:</h1>   
+        <p> Event summary / Description</p> 
+        <p>Locaton</p>
+        <p>start</p>
+        <p>End</p>
+        </div>
+    )
+}
+
+export default CalendarEvents
